Add editing
In delete we did:
0) add controller action , and view button
1) detect event in child Record (onClick: call handleDelete' in child's render)
2) Perform action (sent http request to server, in handleDelete, method in childs props )
3) Notify parent `Records` (handler method through props)
4) Update `Record`'s state

For editing, we use an handleToggle  (1)
But, we need to handle two versions of row: the 'show' version and the 'edit' versions.
We do that in 2 helper methods, 'RecordRow' and RecordForm. We use and extend the table that we already made in the render method.
And add Toggles and input fields where needed.
(Now add update method to controller (0) (the edit button already exist, automagically.
Implement HandleEditRecord method to send request (2)